chore: Update PostgreSQL container configuration
This commit updates the configuration of the PostgreSQL container in the `.devcontainer/compose.yml` file. It adds the `command` option to enable SSL with the specified certificate and key files. Additionally, it corrects the volume paths for the certificate and key files.

The changes in the `.devcontainer/generate-keys.sh` file include setting the ownership of the `server.key` file to `70:70`.